Revert "Add response read timeouts (#10128)"
This reverts commit c4c10aca779f9d94aebf01efd5fb31d4bf25d7ea.
