+++ sdk/core/Azure.Core/CHANGELOG.md
@@ -1,10 +1,8 @@
 # Release History
 
 ## 1.1.0-preview.1 (Unreleased)
-
 ### Fixes and improvements
 - Add OPTIONS and TRACE HTTP request methods.
-- Add `NetworkTimeout` property to `RetryOptions` and apply it to network operations like sending request or reading from the response stream.
 
 ## 1.0.2 (2020-01-10)
