--- sdk/core/Azure.Core/src/Pipeline/Internal/ReadTimeoutStream.cs
@@ -1,128 +1,0 @@
-// Copyright (c) Microsoft Corporation. All rights reserved.
-// Licensed under the MIT License.
-
-using System;
-using System.IO;
-using System.Threading;
-using System.Threading.Tasks;
-
-namespace Azure.Core.Pipeline
-{
-    /// <summary>
-    /// Read-only Stream that will throw a <see cref="OperationCanceledException"/> if it has to wait longer than a configurable timeout to read more data
-    /// </summary>
-    internal class ReadTimeoutStream : ReadOnlyStream
-    {
-        private readonly Stream _stream;
-        private TimeSpan _readTimeout;
-        private CancellationTokenSource _cancellationTokenSource;
-
-        public ReadTimeoutStream(Stream stream, TimeSpan readTimeout)
-        {
-            _stream = stream;
-            _readTimeout = readTimeout;
-            UpdateReadTimeout();
-            _cancellationTokenSource = new CancellationTokenSource();
-        }
-
-        public override int Read(byte[] buffer, int offset, int count)
-        {
-            return _stream.Read(buffer, offset, count);
-        }
-
-        public override async Task<int> ReadAsync(byte[] buffer, int offset, int count, CancellationToken cancellationToken)
-        {
-            var source = StartTimeout(cancellationToken, out bool dispose);
-            try
-            {
-                return await _stream.ReadAsync(buffer, offset, count, source.Token).ConfigureAwait(false);
-            }
-            finally
-            {
-                StopTimeout(source, dispose);
-            }
-        }
-
-        private CancellationTokenSource StartTimeout(CancellationToken additionalToken, out bool dispose)
-        {
-            if (_cancellationTokenSource.IsCancellationRequested)
-            {
-                _cancellationTokenSource = new CancellationTokenSource();
-            }
-
-            CancellationTokenSource source;
-            if (additionalToken.CanBeCanceled)
-            {
-                source = CancellationTokenSource.CreateLinkedTokenSource(additionalToken, _cancellationTokenSource.Token);
-                dispose = true;
-            }
-            else
-            {
-                source = _cancellationTokenSource;
-                dispose = false;
-            }
-
-            _cancellationTokenSource.CancelAfter(_readTimeout);
-
-            return source;
-        }
-
-        private void StopTimeout(CancellationTokenSource source, bool dispose)
-        {
-            _cancellationTokenSource.CancelAfter(Timeout.InfiniteTimeSpan);
-            if (dispose)
-            {
-                source.Dispose();
-            }
-        }
-
-        public override long Seek(long offset, SeekOrigin origin)
-        {
-            return _stream.Seek(offset, origin);
-        }
-
-        public override bool CanRead => _stream.CanRead;
-        public override bool CanSeek => _stream.CanSeek;
-        public override long Length => _stream.Length;
-
-        public override long Position
-        {
-            get => _stream.Position;
-            set => _stream.Position = value;
-        }
-
-        public override int ReadTimeout
-        {
-            get => (int) _readTimeout.TotalMilliseconds;
-            set
-            {
-                _readTimeout = TimeSpan.FromMilliseconds(value);
-                UpdateReadTimeout();
-            }
-        }
-
-        private void UpdateReadTimeout()
-        {
-            try
-            {
-                _stream.ReadTimeout = (int) _readTimeout.TotalMilliseconds;
-            }
-            catch
-            {
-                // ignore
-            }
-        }
-
-        public override void Close()
-        {
-            _stream.Close();
-        }
-
-        protected override void Dispose(bool disposing)
-        {
-            base.Dispose(disposing);
-            _stream.Dispose();
-            _cancellationTokenSource.Dispose();
-        }
-    }
-}
