+++ sdk/core/Azure.Core/src/Shared/RetriableStream.cs
@@ -90,17 +90,14 @@ namespace Azure.Core.Pipeline
                     }
                     catch (Exception e)
                     {
-                        await RetryAsync(e, true, cancellationToken).ConfigureAwait(false);
+                        await RetryAsync(e, true).ConfigureAwait(false);
                     }
                 }
             }
 
-            private async Task RetryAsync(Exception exception, bool async, CancellationToken cancellationToken)
+            private async Task RetryAsync(Exception exception, bool async)
             {
-                bool isNonCustomerCancelledException = exception is OperationCanceledException &&
-                                                    !cancellationToken.IsCancellationRequested;
-
-                if (!_responseClassifier.IsRetriableException(exception) && !isNonCustomerCancelledException)
+                if (!_responseClassifier.IsRetriableException(exception))
                 {
                     ExceptionDispatchInfo.Capture(exception).Throw();
                 }
@@ -135,7 +132,7 @@ namespace Azure.Core.Pipeline
                     }
                     catch (Exception e)
                     {
-                        RetryAsync(e, false, default).EnsureCompleted();
+                        RetryAsync(e, false).EnsureCompleted();
                     }
                 }
             }
