+++ sdk/core/Azure.Core/tests/HttpPipelineFunctionalTests.cs
@@ -2,13 +2,11 @@
 // Licensed under the MIT License.
 
 using System;
-using System.Diagnostics.Tracing;
 using System.IO;
 using System.Net.Http;
 using System.Threading;
 using System.Threading.Tasks;
 using Azure.Core.Pipeline;
-using Microsoft.AspNetCore.Http;
 using NUnit.Framework;
 
 namespace Azure.Core.Tests
@@ -114,14 +112,13 @@ namespace Azure.Core.Tests
         [Test]
         public async Task RetriesTimeoutsServerTimeouts()
         {
-            var testDoneTcs = new CancellationTokenSource();
             int i = 0;
             HttpPipeline httpPipeline = HttpPipelineBuilder.Build(new TestOptions
             {
-                Retry =
+                Transport = new HttpClientTransport(new HttpClient()
                 {
-                    NetworkTimeout = TimeSpan.FromMilliseconds(500)
-                }
+                    Timeout = TimeSpan.FromMilliseconds(500)
+                })
             });
 
             using TestServer testServer = new TestServer(
@@ -129,7 +126,7 @@ namespace Azure.Core.Tests
                 {
                     if (Interlocked.Increment(ref i) == 1)
                     {
-                        await Task.Delay(Timeout.Infinite, testDoneTcs.Token);
+                        await Task.Delay(Timeout.Infinite);
                     }
                     else
                     {
@@ -145,14 +142,11 @@ namespace Azure.Core.Tests
 
             Assert.AreEqual(message.Response.Status, 201);
             Assert.AreEqual(2, i);
-
-            testDoneTcs.Cancel();
         }
 
         [Test]
         public async Task DoesntRetryClientCancellation()
         {
-            var testDoneTcs = new CancellationTokenSource();
             int i = 0;
             HttpPipeline httpPipeline = HttpPipelineBuilder.Build(new TestOptions());
             TaskCompletionSource<object> tcs = new TaskCompletionSource<object>(TaskCreationOptions.RunContinuationsAsynchronously);
@@ -162,7 +156,7 @@ namespace Azure.Core.Tests
                 {
                     Interlocked.Increment(ref i);
                     tcs.SetResult(null);
-                    await Task.Delay(Timeout.Infinite, testDoneTcs.Token);
+                    await Task.Delay(Timeout.Infinite);
                 });
 
             var cts = new CancellationTokenSource();
@@ -179,95 +173,8 @@ namespace Azure.Core.Tests
 
             Assert.ThrowsAsync<TaskCanceledException>(async () => await task);
             Assert.AreEqual(1, i);
-
-            testDoneTcs.Cancel();
         }
 
-#if NETCOREAPP
-        [Test]
-        public async Task RetriesBufferedBodyTimeout()
-        {
-            var testDoneTcs = new CancellationTokenSource();
-            int i = 0;
-            HttpPipeline httpPipeline = HttpPipelineBuilder.Build(new TestOptions
-            {
-                Retry =
-                {
-                    NetworkTimeout = TimeSpan.FromMilliseconds(500)
-                }
-            });
-
-            using TestServer testServer = new TestServer(
-                async context =>
-                {
-                    if (Interlocked.Increment(ref i) == 1)
-                    {
-                        context.Response.StatusCode = 200;
-                        context.Response.Headers.ContentLength = 10;
-                        await context.Response.WriteAsync("1");
-                        await context.Response.Body.FlushAsync();
-
-                        await Task.Delay(Timeout.Infinite, testDoneTcs.Token);
-                    }
-                    else
-                    {
-                        context.Response.StatusCode = 201;
-                        await context.Response.WriteAsync("Hello world!");
-                    }
-                });
-
-            using HttpMessage message = httpPipeline.CreateMessage();
-            message.Request.Uri.Reset(testServer.Address);
-            message.BufferResponse = true;
-
-            await httpPipeline.SendAsync(message, CancellationToken.None);
-
-            Assert.AreEqual(message.Response.Status, 201);
-            Assert.AreEqual("Hello world!", await new StreamReader(message.Response.ContentStream).ReadToEndAsync());
-            Assert.AreEqual(2, i);
-
-            testDoneTcs.Cancel();
-        }
-
-        [Test]
-        public async Task TimeoutsUnbufferedBodyReads()
-        {
-            var testDoneTcs = new CancellationTokenSource();
-
-            HttpPipeline httpPipeline = HttpPipelineBuilder.Build(new TestOptions
-            {
-                Retry =
-                {
-                    NetworkTimeout = TimeSpan.FromMilliseconds(500)
-                }
-            });
-
-            Console.WriteLine($"[{DateTime.Now.ToString("HH:mm:ss")}] Starting");
-            using TestServer testServer = new TestServer(
-                async context =>
-                {
-                    context.Response.StatusCode = 200;
-                    context.Response.Headers.Add("Connection", "close");
-                    await context.Response.WriteAsync("1");
-                    await context.Response.Body.FlushAsync();
-
-                    await Task.Delay(Timeout.Infinite, testDoneTcs.Token);
-                });
-
-            using HttpMessage message = httpPipeline.CreateMessage();
-            message.Request.Uri.Reset(testServer.Address);
-            message.BufferResponse = false;
-            await httpPipeline.SendAsync(message, CancellationToken.None);
-
-             Assert.AreEqual(message.Response.Status, 200);
-             var responseContentStream = message.Response.ContentStream;
-             var buffer = new byte[10];
-             Assert.AreEqual(1, await responseContentStream.ReadAsync(buffer, 0, 1));
-             Assert.That(async () => await responseContentStream.ReadAsync(buffer, 0, 10), Throws.InstanceOf<OperationCanceledException>());
-
-             testDoneTcs.Cancel();
-        }
-#endif
 
         private class TestOptions : ClientOptions
         {
